Move logs to its own controller
