+++ src/Events/Controllers/StorageController.cs
@@ -2,17 +2,14 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 
-using Altinn.Platform.Events.Models;
 using Altinn.Platform.Events.Services.Interfaces;
 
 using CloudNative.CloudEvents;
-
 using Microsoft.ApplicationInsights.DataContracts;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-
 using Swashbuckle.AspNetCore.Annotations;
 
 namespace Altinn.Platform.Events.Controllers
@@ -26,7 +23,6 @@ namespace Altinn.Platform.Events.Controllers
     public class StorageController : ControllerBase
     {
         private readonly IEventsService _eventsService;
-        private readonly ITraceLogService _traceLogService;
         private readonly ILogger _logger;
 
         /// <summary>
@@ -34,12 +30,10 @@ namespace Altinn.Platform.Events.Controllers
         /// </summary>
         public StorageController(
             IEventsService eventsService,
-            ITraceLogService traceLogService,
             ILogger<StorageController> logger)
         {
             _logger = logger;
             _eventsService = eventsService;
-            _traceLogService = traceLogService;
         }
 
         /// <summary>
@@ -67,29 +61,6 @@ namespace Altinn.Platform.Events.Controllers
                 _logger.LogError(e, "Temporarily unable to save cloudEventId {CloudEventId} to storage, please try again.", cloudEvent?.Id);
                 return StatusCode(503, e.Message);
             }
-        }
-
-        /// <summary>
-        /// Create a new trace log for cloud event with a status code.
-        /// </summary>
-        /// <param name="logEntry">The event wrapper associated with the event for logging <see cref="CloudEventEnvelope"/></param>
-        /// <returns></returns>
-        [Authorize(Policy = "PlatformAccess")]
-        [HttpPost("logs")]
-        [Consumes("application/json")]
-        [SwaggerResponse(201, Type = typeof(Guid))]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<IActionResult> Logs([FromBody] LogEntryDto logEntry)
-        {
-            var result = await _traceLogService.CreateWebhookResponseEntry(logEntry);
-
-            if (string.IsNullOrEmpty(result))
-            {
-                return BadRequest();
-            }
-
-            return Created();
         }
 
         [ExcludeFromCodeCoverage]
